refactor: app 모듈에 pricealert 패키지 생성 및 파일 구조 정리  - #86
